simple_dify_on_wechat/requirements.txt
requests
websocket-client
python-dotenv